==> 19-paper.Rmd <==
# Presentation and Paper {#final}

## Presentation

### Presentation Procedure: 

1. Everyone accesses to the course Zoom account

2. Prepare presentation slides on your own PC.

3. Each person gives a presentation according to the list. (Each person has 7 minutes).

4. Presenters project their presentation slides onto the Zoom screen using the screen sharing function. 

5. Presenter unmutes and delivers the presentation. 

6. The next presenter prepares his/her slides and waits.

### Submission: Notebook file and the Presentation file

Please submit the following.

Notebook files and presentation slides used in the presentation (if the presentation is given in R Notebook, no slides are required).

Deadline: 23:59 of February 25. 

## Final Report

### Submission of a Final Report

Students are required to submit a final report on a research topic.

1. Students select a research topic.
2. Collecting data on your research topic.
3. Analyze data using methods learnt in lectures.
4. Discuss about your findings.

The format of the report is not specified, but the following is an example.

Content of a final report:

1. Title of your research
2. Abstract
3. Data
4. Analysing data
5. Findings
6. Discussion
7. References.

Students are requested to submit a paper, with a supporting document in R Notebook to prepare your essay. Submit the PDF file of the final report and the R Notebbok file on Moodle.

Length: less than 8 pages in A4 format including figures and tables.　Upload it on Moodle. 

Deadline of submission: March 08. 

## Topics Chosen by Students

1. Migration and Economic performance

2. Education Costs and Fertility Rate in Japan

3. Analysis of Athletes Nations Distribution and Gender Equality in the Olympics

4. The correlation between the trade surplus and the economic development

5. People’s views on Religion, Nature, and Science

6. The rate of patent applications in scientific /technical publications from 1985 -2015

7. The Correlation between Democracy and Life Satisfaction by Looking at the OWID Data

8. Impact of entrepreneurship on economic growth; Case of Pakistan & India

9. Environmental Impacts of Food and Agriculture

10. Relationship between Government expenditure on education, Democracy and GDP per capita in Sri Lanka

11. Relationship Between Climate Change and Agricultural Production in Pakistan


12. The Effect of ICT Investment to the Economic Growth: Case Study in OECD

13. Impact of Personal Remittances & Exchange rates on Gross Savings Rate in Pakistan

14. Democracy & Trust in the Government

15. World Population Growth

16. The Development of Renewable Energy Internationally 

17. Poverty Measurement and Culture in Israel

18. Relationship between war violence and happiness in the world
